Use stripped-down CBaseMerkleTx for auxpow.
Previously, all code for CMerkleTx was moved from upstream's wallet/wallet
file to auxpow, because CAuxPow depends on CMerkleTx.  This was problemtic,
because the code actually depends on validation, which is not in the
"common" library (as auxpow is).

This change splits CMerkleTx into a CBaseMerkleTx that is the part
required for CAuxPow, basically including the member fields and
serialisation.  The other things, including the methods used by the
wallet that depend on validation, are only added in wallet/wallet to
CMerkleTx that extends the new CBaseMerkleTx.

In addition to solving the invalid dependency issue, this also moves some
code back to where it is upstream, reducing merge conflicts.

See https://github.com/namecoin/namecoin-core/issues/231 for more details.
