+++ src/wallet/wallet.h
@@ -7,7 +7,7 @@
 #define BITCOIN_WALLET_WALLET_H
 
 #include <amount.h>
-#include <auxpow.h> // contains CMerkleTx
+#include <auxpow.h> // contains CBaseMerkleTx
 #include <policy/feerate.h>
 #include <streams.h>
 #include <tinyformat.h>
@@ -206,6 +206,40 @@ struct COutputEntry
     CTxDestination destination;
     CAmount amount;
     int vout;
+};
+
+/** A transaction with a merkle branch linking it to the block chain. */
+class CMerkleTx : public CBaseMerkleTx
+{
+private:
+  /** Constant used in hashBlock to indicate tx has been abandoned */
+    static const uint256 ABANDON_HASH;
+
+public:
+
+    CMerkleTx() = default;
+
+    explicit CMerkleTx(CTransactionRef arg)
+      : CBaseMerkleTx(arg)
+    {}
+
+    void SetMerkleBranch(const CBlockIndex* pindex, int posInBlock);
+
+    /**
+     * Return depth of transaction in blockchain:
+     * <0  : conflicts with a transaction this deep in the blockchain
+     *  0  : in memory pool, waiting to be included in a block
+     * >=1 : this many blocks deep in the main chain
+     */
+    int GetDepthInMainChain(const CBlockIndex* &pindexRet) const;
+    int GetDepthInMainChain() const { const CBlockIndex *pindexRet; return GetDepthInMainChain(pindexRet); }
+    bool IsInMainChain() const { const CBlockIndex *pindexRet; return GetDepthInMainChain(pindexRet) > 0; }
+    int GetBlocksToMaturity() const;
+    bool hashUnset() const { return (hashBlock.IsNull() || hashBlock == ABANDON_HASH); }
+    bool isAbandoned() const { return (hashBlock == ABANDON_HASH); }
+    void setAbandoned() { hashBlock = ABANDON_HASH; }
+
+    bool IsCoinBase() const { return tx->IsCoinBase(); }
 };
 
 //Get the marginal bytes of spending the specified output
